+++ frontend/src/api/user.js
@@ -46,6 +46,14 @@ export function updateUser(data) {
     })
 }
 
+export function resetUserPassword(data) {
+    return request({
+        url: '/user/reset-password',
+        method: 'put',
+        data
+    })
+}
+
 export function deleteUser(id) {
     return request({
         url: `/user/delete/${id}`,